move Main in components
