+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import NavBar from './components/NavBar/NavBar';
-import { Main } from './containers/Main/Main';
+import { Main } from './components/Main/Main';
 import Cart from './components/Cart/Cart';
 import Description from './components/Description/Description';
 import { catalog } from './utilities/catalog';
