+++ src/components/Cart/Cart.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import './Cart.css';
-import { currentDiscount, currentDiscountedElements } from '../../containers/Main/Main';
+import { currentDiscount, currentDiscountedElements } from '../Main/Main';
 import { catalog } from '../../utilities/catalog';
 import CartButtons from '../Buttons/CartButtons/CartButtons';
 import Cards from '../Card/Cards/Cards'
