+++ src/components/Main/Main.css
@@ -52,10 +52,11 @@ body {
 
 .main-container .description-container {
   width: 80%;
-  height: 100%;
+  height: auto;
   margin: 50px auto;
   position: relative;
   margin-top: 80px;
+  padding: 10px;
 }
 
 .main-container .description-container h1 {
@@ -73,7 +74,7 @@ body {
   font-size: 3.2em;
   font-size: #444953;
   color: inherit;
-  font-size: 2.0em;
+  font-size: 2em;
 }
 
 .main-container .description-container p {
