+++ src/components/Main/Main.js
@@ -1,19 +1,16 @@
 import React from 'react';
 import './Main.css';
-import { DisplayCatalog, currentDiscountedElements, currentDiscount } from '../../components/DisplayCatalog/DisplayCatalog';
-import { FaTh, FaThLarge } from 'react-icons/fa';
+import { DisplayCatalog, currentDiscountedElements, currentDiscount } from '../DisplayCatalog/DisplayCatalog';
+import MainButtons from '../Buttons/MainButtons/MainButtons'
 
 const Main = props => (
     <div class='main-container'>
         <div className='description-container'>
-            <h1>Hello</h1>
+            <h1>Books for life</h1>
             <h2>Explore our range of books</h2>
         </div>
         <div id='books-container' >
-            <div className='display-wrapper'>
-                <FaTh id='normal-display' onClick={props.displayProducts.bind(this, 'normal')} title='default display' />
-                <FaThLarge id='large-display' onClick={props.displayProducts.bind(this, 'large')} title='large display' />
-            </div>
+            <MainButtons displayProducts={props.displayProducts} />
             <DisplayCatalog
                 display={props.display}
                 showDescription={props.showDescription}
