+++ src/components/Main/Main.scss
@@ -1,144 +1,38 @@
-@import '../../style/mixins';
-@import '../../style/colors';
-@import '../../style/layout-mixin';
+@import "../../style/mixins";
+@import "../../style/colors";
+@import "../../style/layout-mixin";
 
 body {
     overflow: hidden auto;
 }
 
 .main-container {
-    // @include build-box(100%, auto, none, 0);
     @include box(100%, auto, 0);
+
     .description-container {
-    // @include build-box(80%, 100%, none, 50px auto, 10px);
-    @include box(80%, 100%, 50px auto, relative);
-    margin-top: 80px;
-      h1 {
-        @include font(header, initial, $oranged-yellow);
-        font-size: 3.8em;
-    }
-     h2 {
-        @include font(header, $dark-grayish);
-        font-size: 2.0em;
-    }
-    p {
-        @include font(paragraph);
-        margin-top: 10px;
-    }
+        @include box(80%, auto, 50px auto, relative);
+        margin-top: 80px;
+        padding: 10px;
+        h1 {
+            @include font(header, initial, $oranged-yellow);
+            font-size: 3.8em;
+        }
+        h2 {
+            @include font(header, $dark-grayish);
+            font-size: 2em;
+        }
+        p {
+            @include font(paragraph);
+            margin-top: 10px;
+        }
     }
 
-    #books-container  {
+    #books-container {
         @include box(80%, 60%, 10px auto, relative);
         display: flex;
         justify-content: space-between;
         flex-wrap: wrap;
-        border: {
-            left: 2px solid $shaded-purple;
-            right: 2px solid $shaded-purple;
-        }
+        border-left: 2px solid $shaded-purple;
+        border-right: 2px solid $shaded-purple;
     }
 }
-
-        // //style layout buttons
-        // .display-wrapper {
-        //     position: absolute;
-        //     top: 0;
-        //     right: 5px;
-        //     #normal-display, #large-display {
-        //       font-size: 3em;
-        //       color:$dark-grayish;
-        //       margin: 1px 3px;
-        //     }
-        //     #normal-display:hover, #large-display:hover {
-        //       cursor: pointer;
-        //       color: gray;
-        //     }
-        //   }
-
-
-
-//     //outsource below to DisplayCatalog
-//       .card {
-//        @include build-box(300px, 400px, none, 10px 0, 10px);
-//        display: inline-block;   
-//        img {
-//         width: 70%;
-//         height: 83.5%;
-//         display: block;
-//         margin: 0 auto;
-//         object-fit: cover;
-//         margin-top: 10px;
-//        }
-//         h2 {
-//             @include font(header, 2.2em, $shaded-purple);
-//         }
-//         h3 {
-//             @include font(header, 1.8em, #3F3937);
-            
-//         }
-//            p {
-//                @include font(header, 1.7em);
-//                font-weight: bold;
-//             }
-
-//             .discountedPrice {
-//                 @include font(discountedPrice);
-//             }
-//             #add-btn-main {
-//                 margin-left: 10px;
-//                 font-size: 1.7em;
-//                 color: $oranged-yellow;
-//                 opacity: .8;
-//             }
-//             #add-btn-main:hover {
-//                 cursor: pointer;
-//                 opacity: 1;
-                
-//             }
-
-//       }
-//       .card.large{
-//         @include build-box(400px, 500px, none, 10px 20px, 10px);
-//         h2 {
-//             font-size: 2.6em;
-//         }
-//         h3 {
-//             font-size: 2em;
-//         }
-//            p {
-//                font-size: 1.9em;
-//             }
-//             #add-btn-main {
-//                 font-size: 3rem;
-//             }
-//     }
-//     } 
-//     img:hover {
-//         cursor: pointer;
-//             box-shadow: 0px 2px 1px rgba(107, 103, 103, .5), 2px 0px 1px rgba(107, 103, 103, .5), 
-//             0px -2px 1px rgba(107, 103, 103, 0.5), -2px -2px 1px rgba(107, 103, 103, .5);
-//     }
-    
-// }
-
-// @media screen and (max-width: 1040px){ 
-//     .main-container #books-container .card.large {
-//         margin: 10px auto;
-//         img {
-//             min-width: 200px;
-//         }
-//     } 
-// }
-
-// @media screen and (max-width: 808px){
-//     .main-container #books-container .card {
-//         margin: 10px auto;
-//     } 
-// }
-
-// @media screen and (max-width: 460px){
-//     .main-container #books-container .card.large {
-//         width: 300px;
-//         height: 400px;
-//     } 
-// }
