--- src/containers/Main/Main.css
@@ -1,104 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Noto+Sans&display=swap");
-/*
-  Text types: 
-     
-   > Book title (-Main + Cart-)
-   > Author (-Main + Cart-)
-   > price (Main)
-   > discounted price
-   > buttons (-Main + Cart-)
-   > cart buttons
-
-
-*/
-/*
- ORDER for properties
-  1) width/height
-  2) display (and other related rules)
-  3) margin/padding
-  4) background
-  5) border
-  6) children
-*/
-/*
- ORDER for properties
- ! mixins
-  1) width/height
-  2) display (and other related rules)
-  2.5) position (and other related rules)
-  3) margin/padding
-  4) background
-  5) border
-  6) color
-  7) children
-
-  Text
-
-  margin/padding
-  font
-  text
-  color
-*/
-body {
-  overflow: hidden auto;
-}
-
-.main-container {
-  width: 100%;
-  height: auto;
-  margin: 0;
-  position: initial;
-}
-
-.main-container .description-container {
-  width: 80%;
-  height: auto;
-  margin: 50px auto;
-  position: relative;
-  margin-top: 80px;
-  padding: 10px;
-}
-
-.main-container .description-container h1 {
-  font-family: 'Noto Sans', sans-serif, serif, monospace;
-  text-align: center;
-  font-size: 3.2em;
-  font-size: initial;
-  color: #FFA100;
-  font-size: 3.8em;
-}
-
-.main-container .description-container h2 {
-  font-family: 'Noto Sans', sans-serif, serif, monospace;
-  text-align: center;
-  font-size: 3.2em;
-  font-size: #444953;
-  color: inherit;
-  font-size: 2em;
-}
-
-.main-container .description-container p {
-  font-family: 'Noto Sans', sans-serif, serif, monospace;
-  font-size: 2em;
-  line-height: 1.5;
-  color: inherit;
-  margin-top: 10px;
-}
-
-.main-container #books-container {
-  width: 80%;
-  height: 60%;
-  margin: 10px auto;
-  position: relative;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-pack: justify;
-      -ms-flex-pack: justify;
-          justify-content: space-between;
-  -ms-flex-wrap: wrap;
-      flex-wrap: wrap;
-  border-left: 2px solid #673F4C;
-  border-right: 2px solid #673F4C;
-}
-/*# sourceMappingURL=Main.css.map */
