--- src/containers/Main/Main.scss
@@ -1,38 +1,0 @@
-@import "../../style/mixins";
-@import "../../style/colors";
-@import "../../style/layout-mixin";
-
-body {
-    overflow: hidden auto;
-}
-
-.main-container {
-    @include box(100%, auto, 0);
-
-    .description-container {
-        @include box(80%, auto, 50px auto, relative);
-        margin-top: 80px;
-        padding: 10px;
-        h1 {
-            @include font(header, initial, $oranged-yellow);
-            font-size: 3.8em;
-        }
-        h2 {
-            @include font(header, $dark-grayish);
-            font-size: 2em;
-        }
-        p {
-            @include font(paragraph);
-            margin-top: 10px;
-        }
-    }
-
-    #books-container {
-        @include box(80%, 60%, 10px auto, relative);
-        display: flex;
-        justify-content: space-between;
-        flex-wrap: wrap;
-        border-left: 2px solid $shaded-purple;
-        border-right: 2px solid $shaded-purple;
-    }
-}
